ICD inherit CDM
