--- EduCDM/ICD/ICD/_search_space.json
@@ -1,3 +1,0 @@
-{
-  "beta": {"_type":"choice","_value":[0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 0.95, 0.99]}
-}
