--- EduCDM/ICD/ICD/sym/__init__.py
@@ -1,5 +1,0 @@
-# coding: utf-8
-# 2022/1/29 @ tongshiwei
-
-from .net import get_net, get_loss, ICD, DualICD, get_dual_loss
-from .fit_eval import fit_f, eval_f, dual_fit_f, stableness_eval, turning_point
