+++ EduCDM/ICD/constant.py
@@ -1,1 +1,2 @@
-path_prefix = '/home/yutingh/EduCDM/ICD/'
+path_prefix = '/home/yutingh/EduCDM/ICD'
+path_prefix = '/Users/vivianhong/Documents/py/EduCDM/EduCDM/ICD'
