+++ EduCDM/ICD/sym/net/net.py
